BUGFIX: Apply MariaDB specific fix only for MariaDB database

+++ Migrations/Mysql/Version20150224171107.php
@@ -13,7 +13,7 @@ class Version20150224171107 extends AbstractMigration
      * @param Schema $schema
      * @return void
      */
-    public function up(Schema $schema): void 
+    public function up(Schema $schema): void
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != "mysql");
 
@@ -23,11 +23,11 @@ class Version20150224171107 extends AbstractMigration
         $this->addSql("ALTER TABLE typo3_neos_eventlog_domain_model_event DROP persistence_object_identifier, CHANGE parentevent parentevent INT UNSIGNED DEFAULT NULL, CHANGE uid uid INT UNSIGNED AUTO_INCREMENT NOT NULL");
         $this->addSql("ALTER TABLE typo3_neos_eventlog_domain_model_event ADD CONSTRAINT FK_30AB3A75B684C08 FOREIGN KEY (parentevent) REFERENCES typo3_neos_eventlog_domain_model_event (uid)");
         $indexes = $this->sm->listTableIndexes('typo3_neos_eventlog_domain_model_event');
-        if (array_key_exists('uid', $indexes)) {
+        if (array_key_exists('uid', $indexes) && $this->connection->getDatabasePlatform() instanceof MariaDBPlatform) {
             $this->addSql("ALTER TABLE typo3_neos_eventlog_domain_model_event DROP INDEX uid, ADD INDEX olduid (uid)");
         }
         $this->addSql("ALTER TABLE typo3_neos_eventlog_domain_model_event ADD PRIMARY KEY (uid)");
-        if (array_key_exists('uid', $indexes)) {
+        if (array_key_exists('uid', $indexes) && $this->connection->getDatabasePlatform() instanceof MariaDBPlatform) {
             $this->addSql("ALTER TABLE typo3_neos_eventlog_domain_model_event DROP INDEX olduid");
         }
     }
@@ -36,7 +36,7 @@ class Version20150224171107 extends AbstractMigration
      * @param Schema $schema
      * @return void
      */
-    public function down(Schema $schema): void 
+    public function down(Schema $schema): void
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != "mysql");
